clean up the test output

+++ flow/tests/testthat/test-wrap-rglr.R
@@ -118,19 +118,22 @@ test_that("Unit test of wrap.rglr.R", {
   
   #1 Test 1
   
-   wrap.rglr(
+  wrap.rglr(
     DirIn = DirIn,
     DirOutBase = DirOutBase,
     DirSubCopy = "data_flags",
     ParaRglr = ParaRglr
   )
   
-  # remove symbolic link
+  # remove the test output symbolic link
   DirSrc = 'CFGLOC101670'
-  cmdLs <- base::paste0('ls ',base::paste0(DirSrc))
-  cmdSymbLink <- base::paste0('rm ',base::paste0(DirSrc))
+  cmdLs <- base::paste0('ls ', base::paste0(DirSrc))
+  exstDirSrc <- base::unlist(base::lapply(DirSrc, base::dir.exists))
   
-  rmSymbLink <- base::lapply(cmdSymbLink,base::system) 
+  if (exstDirSrc) {
+    cmdSymbLink <- base::paste0('rm ', base::paste0(DirSrc))
+    rmSymbLink <- base::lapply(cmdSymbLink, base::system)
+  }
   
   dirInData <- base::paste0(DirIn, '/data')
   dirInFlags <- base::paste0(DirIn, '/flags')
@@ -195,4 +198,4 @@ test_that("Unit test of wrap.rglr.R", {
                                   DirOutBase = DirOutBase,
                                   ParaRglr = ParaRglr_NA),silent = TRUE)
   
-  })
+})
